Moved Icon into atoms directory
