+++ src/components/FloatingSearch.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import { useRouter } from "next/router";
-import { Icon } from "./icon/Icon";
+import { Icon } from "./atoms/icon/Icon";
 import { Divider } from "./dividers/Divider";
 import { CleanRouterQuery } from "@utils/cleanRouterQuery";
 import { QUERY_PARAMS } from "@constants/queryParams";
