+++ src/components/documents/EmptyPassages.tsx
@@ -1,4 +1,4 @@
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 
 type TProps = {
   hasQueryString: boolean;
