+++ src/components/forms/SearchDropdown.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import useConfig from "@hooks/useConfig";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 import { QUERY_PARAMS } from "@constants/queryParams";
 import { TGeography } from "@types";
 import { systemGeoCodes } from "@constants/systemGeos";
