+++ src/components/map/GeographySelect.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import SuggestList from "@components/filters/SuggestList";
 import { sortData } from "@utils/sorting";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 
 interface ByTextInputProps {
   title: string;
