+++ src/components/search/SearchResult.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@components/atoms/button/Button";
 import { FamilyListItem } from "@components/document/FamilyListItem";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 import { ToolTipSSR } from "@components/tooltip/TooltipSSR";
 import { MAX_RESULTS } from "@constants/paging";
 import { TMatchedFamily } from "@types";
