+++ themes/cclw/components/Instructions.tsx
@@ -5,7 +5,7 @@ import useConfig from "@hooks/useConfig";
 import { calculateTotalFamilies } from "@helpers/getFamilyCounts";
 
 import { Button } from "@components/atoms/button/Button";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 
 import { INSTRUCTIONS } from "@cclw/constants/instructions";
