+++ themes/cclw/components/LandingSearchForm.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useRef, ChangeEvent } from "react";
 
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 import { SearchDropdown } from "@components/forms/SearchDropdown";
 import { Button } from "@components/atoms/button/Button";
