+++ themes/cclw/components/Menu.tsx
@@ -1,6 +1,6 @@
 import { useState, useRef } from "react";
 import useOutsideAlerter from "@hooks/useOutsideAlerter";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 import DropdownMenuItem from "@components/menus/DropdownMenuItem";
 import DropdownMenuWrapper from "@components/menus/DropdownMenuWrapper";
 import MENU_LINKS from "@cclw/constants/menuLinks";
