+++ themes/cpr/components/Summary.tsx
@@ -1,7 +1,7 @@
 import { SiteWidth } from "@components/panels/SiteWidth";
 
 import { ExternalLink } from "@components/ExternalLink";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 import { Heading } from "@components/typography/Heading";
 
 const Summary = () => {
