+++ themes/mcf/components/FooterComponents/socialMediaData.tsx
@@ -1,4 +1,4 @@
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 
 import { colors } from "@mcf/constants/colors";
