+++ themes/mcf/components/HeaderComponents/Menu.tsx
@@ -1,7 +1,7 @@
 import { useState, useRef, useCallback } from "react";
 
 import useOutsideAlerter from "@hooks/useOutsideAlerter";
-import { Icon } from "@components/icon/Icon";
+import { Icon } from "@components/atoms/icon/Icon";
 import DropdownMenuItem from "@components/menus/DropdownMenuItem";
 import DropdownMenuWrapper from "@components/menus/DropdownMenuWrapper";
